Fix gen-for-case and case-arm-block
Matching of the { } had gone wrong, especially with regard to escape
sequences. Inside quoted strings which are actually regexes one has
to escape the \ which escapes the { for example, so to match a { symbol
in a quoted string you actually need to put \\{ in the quoted regex!

(Quotes are required around regexes containing colons, since otherwise
they mess up the yaml parsing.)