Allow comments to contain semicolons
Previously comments inside functions which had a ; in the comment string
and were after another statement were highlighed as statements of various
sorts up to the ; since the regexes that were matching the statement
contents up to the first ; on the line were being greedy not reluctant.

Need to reorder consideration of statements to make this work though.